development/tke-morning-intelligence/services/content-engine/pyproject.toml
[project]
name = "tke-morning-intelligence-content-engine"
version = "1.0.0"
description = "Content generation engine for TKE Morning Intelligence using Vertex AI Gemini 3.x"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "google-genai>=1.0.0",
    "google-cloud-firestore>=2.19.0",
    "pydantic>=2.10.0",
    "httpx>=0.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
